Quality control flags

Quality control flags

Details

At the moment there are 29 QC flags, numbered from 1 to 30 with disabled flag (qc 8, 9 and 20).

1. OBIS data format: are the required fields from the OBIS Schema completed?

2. Taxonomy: is the taxon name matched to WoRMS?

3. Taxonomy: is the taxon level genus or lower?

4. Geography (lat/lon): are the latitude/longitude values different from zero?

5. Geography (lat/lon): are the latitude/longitude values within their possible boundaries? (world coordinates)

6. Geography (lat/lon): are the coordinates situated in sea or along the coastline (20 km buffer)?

7. Completeness (date/time): is the sampling year (start/end) completed and valid?

10. OBIS data format: is the 'Basis of Record' documented, and is an existing OBIS code used?

11. Completeness (date/time): is the sampling date (year/month/day; start/end) valid?

12. Completeness (date/time): if a start and end date are given, is the start before the end?

13. Completeness (date/time): if a sampling time is given, is this valid and is the time zone completed?

14. Completeness (presence/abundance/biomass): is the value of the field 'ObservedIndividualCount' empty or >0?

15. Completeness (presence/abundance/biomass): is the value of the field 'Observedweight' empty or >0?

16. Completeness (presence/abundance/biomass): is the field 'SampleSize' completed if the field 'ObservedIndividualCount' is >0?

17. OBIS data format: is the value of the field 'Sex' empty or is an existing OBIS code used?

18. Geography (depth): is minimum depth <= maximum depth?

19. Geography (depth): is the sampling depth possible when compared with GEBCO depth map (incl. margin)?

21. Species outliers (environment/depth): is the observation within six MADs from the median depth of this taxon?

22. Species outliers (environment/depth): is the observation within three IQRs from the first & third quartile depth of this taxon?

23. Species outliers (environment/SSS): is the observation within six MADs from the median sea surface salinity (SSS) of this taxon?

24. Species outliers (environment/SSS): is the observation within three IQRs from the first & third quartile sea surface salinity (SSS) of this taxon?

25. Species outliers (environment/SST): is the observation within six MADs from the median sea surface temperature (SST) of this taxon?

26. Species outliers (environment/SST): is the observation within three IQRs from the first & third quartile sea surface temperature (SST) of this taxon?

27. Species outliers (geography): is the observation within six MADs from the distance to the geographic centroid of this taxon?

28. Species outliers (geography): is the observation within three IQRs from the first & third quartile distance to the geographic centroid of this taxon?

29. Dataset outliers (geography): is the observation within six MADs from the distance to the geographic centroid of this dataset?

30. Dataset outliers (geography): is the observation within three IQRs from the first & third quartile distance to the geographic centroid of this dataset?

References

Vandepitte, L., Bosch, S., Tyberghein, L., Waumans, F., Vanhoorne, B., Hernandez, F., ??? Mees, J. (2015). Fishing for data and sorting the catch: assessing the data quality, completeness and fitness for use of data in marine biogeographic databases. Database, 2015, bau125???bau125. doi:10.1093/database/bau125

See also

qcflags