Find occurrences.

Find occurrences.

occurrence(scientificname = NULL, year = NULL, obisid = NULL, aphiaid = NULL,
  groupid = NULL, resourceid = NULL, nodeid = NULL, areaid = NULL, startdate = NULL,
  enddate = NULL, startdepth = NULL, enddepth = NULL, geometry = NULL, qc = NULL,
  fields = NULL, verbose = FALSE)

Arguments

scientificname
character vector. The full scientific name, with authorship and date information if known.

year
integer vector. The year in which the Event occurred.

obisid
integer vector. The OBIS identifier of the species.

aphiaid
integer vector. The WoRMS identifier of the species.

groupid
integer. The taxonomic group id. See also group for the list of taxonomic groups.

resourceid
integer vector. The dataset identifier. See also dataset for querying the list of datasets.

nodeid
integer vector. The OBIS node identifier. See also node for the list of nodes.

areaid
integer vector. The OBIS area identifier (country, marine world heritage site, ABNJ, EBSA, ...). See also area for the list of areas.

startdate
The earliest date on which the Event occurred.

enddate
The latest date on which the Event occurred.

startdepth
The minimum depth below the sea surface.

enddepth
The maximum depth below the sea surface.

geometry
A wkt geometry string.

qc
A vector of quality control flags you want to filter on. List of QC flags.

fields
A vector of field names you want to have returned in order, by default all fields with values are returned.

verbose
logical. Optional parameter to enable verbose logging (default = FALSE).

Value

The occurrence records.

See also

leafletmap qcflags checklist dataset area node group

Examples

records <- occurrence(scientificname = "Abra sibogai")

#> Retrieved 12 records of 12 (100%)

records <- occurrence(aphiaid = 141438, startdate = as.Date("2007-10-10"))

#> Retrieved 121 records of 121 (100%)

records <- occurrence(aphiaid = 141438, geometry = "POLYGON ((0 0, 0 45, 45 45, 45 0, 0 0))")

#> Retrieved 378 records of 378 (100%)

records <- occurrence(scientificname = "Abra sibogai", qc = c(1:6, 27))

#> Retrieved 8 records of 8 (100%)

records <- occurrence(scientificname = "Abra sibogai",
                      fields = c("species", "decimalLongitude", "decimalLatitude"))

#> Retrieved 12 records of 12 (100%)